Autoscale over unmanaged instance groups
This allows an admin to set up an unmanaged instance group and then
manage autoscaling over the nodes in that group. Instead of deleting or
creating instances, the autoscaler stops or starts those instances. The
target count is measured as the number of currently running instances.

Common scenarios where this is useful include preconfigured nodes that
have externally injected secrets or configuration that cannot be
automated within a managed instance group. It's also a model that other
clouds may leverage.

+++ cluster-autoscaler/cloudprovider/gce/gce_cloud_provider.go
@@ -193,6 +193,7 @@ type Mig struct {
 	minSize         int
 	maxSize         int
 	autoprovisioned bool
+	unmanaged       bool
 	exist           bool
 	nodePoolName    string
 	spec            *autoprovisioningSpec
@@ -296,6 +297,9 @@ func (mig *Mig) DeleteNodes(nodes []*apiv1.Node) error {
 		}
 		refs = append(refs, gceref)
 	}
+	if mig.unmanaged {
+		return mig.gceManager.StopInstances(refs)
+	}
 	return mig.gceManager.DeleteInstances(refs)
 }
 
@@ -385,6 +389,9 @@ func buildMig(value string, gceManager *GceManager) (*Mig, error) {
 
 	if mig.Project, mig.Zone, mig.Name, err = ParseMigUrl(spec.Name); err != nil {
 		return nil, fmt.Errorf("failed to parse mig url: %s got error: %v", spec.Name, err)
+	}
+	if _, err := gceManager.gceService.InstanceGroupManagers.Get(mig.Project, mig.Zone, mig.Name).Do(); err != nil {
+		mig.unmanaged = true
 	}
 	return &mig, nil
 }
